# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "TES SWAGGER ULBI",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/indrariksa",
            "email": "indra@ulbi.ac.id"
        },
        "version": "1.0"
    },
    "host": "pbo-tugasbesar-c553f4142173.herokuapp.com",
    "basePath": "/",
    "paths": {
        "/api/mahasiswa": {
            "get": {
                "description": "Mengambil semua data mahasiswa",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Get All Data.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mahasiswa.DataMahasiswa"
                        }
                    }
                }
            },
            "post": {
                "description": "Input data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Insert data.",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mahasiswa.DataMahasiswa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mahasiswa.DataMahasiswa"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/mahasiswa/{id}": {
            "get": {
                "description": "Ambil per ID data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Get By ID Data.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Masukan Nama",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mahasiswa.DataMahasiswa"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Ubah data presensi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Update data presensi.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Masukan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mahasiswa.DataMahasiswa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mahasiswa.DataMahasiswa"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "mahasiswa.DataMahasiswa": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "asalsekolah": {
                    "type": "string"
                },
                "namalengkap": {
                    "type": "string"
                },
                "prodi": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        }
    }
}